DevreotesLabResearchChatbot/requirements.txt
neo4j
pymupdf
pytesseract
Pillow
scispacy
langchain
langchain-core
langchain-community
langchain-openai
sentence-transformers
requests
python-dotenv
gradio
fastapi
uvicorn[standard]
